queue now UInt64 Mouse move added

+++ bochs/gui/cocoa_device.h
@@ -56,7 +56,7 @@
     void set_font(bool font2, unsigned pos, unsigned char *charmap);
     void draw_char(bool crsr, bool font2, unsigned char fgcolor, unsigned char bgcolor, unsigned short int charpos, unsigned short int x, unsigned short int y, unsigned char w, unsigned char h);
     bool hasKeyEvent();
-    unsigned getKeyEvent();
+    unsigned long getKeyEvent();
     void graphics_tile_update(unsigned char *tile, unsigned x, unsigned y, unsigned w, unsigned h);
 
   };
